Able to create trip in interface
