+++ frontend/components/map/SaveTripProcess/StepsStatusBar.tsx
@@ -18,25 +18,7 @@ const StepsWrapper = styled.div`
   left: 3%;
   top: 22%;
 `;
-const StepButtonGroup = styled.div`
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-`;
-const ExitBtn = styled(Button)`
-  display: ${(props: {step: number}) =>
-    props.step === 0 ? "flex" : "none"};
-`;
-const PreviousBtn = styled(Button)`
-  display: ${(props: {step: number}) =>
-    props.step > 0 ? "flex" : "none"};
-`;
-const NextBtn = styled(Button)`
-  display: flex;
-`;
-const DoneBtn = styled(Button)`
-  display: flex;
-`;
+
 const PreviewImage = styled.img`
   width: 200px;
   height: 200px;
@@ -79,53 +61,7 @@ const StepsStatusBar: React.SFC<Props> = ({
               alt="Google Map Preview Image"
             />
           ) : null}
-          <WrappedFinishTripForm
-            step={step}
-            buttonGroup={
-              <StepButtonGroup>
-                <ExitBtn
-                  step={step}
-                  type="danger"
-                  onClick={() => setStep(-1)}
-                >
-                  Exit
-                </ExitBtn>
-                <PreviousBtn
-                  step={step}
-                  onClick={() =>
-                    setStep((prevState: number) => prevState - 1)
-                  }
-                  disabled={step === 0 ? true : false}
-                >
-                  Previous
-                </PreviousBtn>
-                {step === 2 ? null : (
-                  <NextBtn
-                    step={step}
-                    type="primary"
-                    onClick={() =>
-                      setStep((prevState: number) => prevState + 1)
-                    }
-                    disabled={step > 2 ? true : false}
-                  >
-                    Next
-                  </NextBtn>
-                )}
-                {step === 2 ? (
-                  <DoneBtn
-                    step={step}
-                    type="primary"
-                    onClick={() =>
-                      setStep((prevState: number) => prevState + 1)
-                    }
-                    disabled={step > 2 ? true : false}
-                  >
-                    Done
-                  </DoneBtn>
-                ) : null}
-              </StepButtonGroup>
-            }
-          />
+          <WrappedFinishTripForm step={step} setStep={setStep} />
         </Card>
       </StepsWrapper>
     );
